% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lm.R
\name{sim_lm}
\alias{sim_lm}
\title{Simulate from a Linear Model}
\usage{
sim_lm(
  n = 25,
  each = 1,
  b0 = list(b0 = 0),
  bj = list(b1 = 0, b2 = 0),
  distj = list(rnorm, rnorm),
  paramsj = list(c(0, 0), c(1, 1)),
  m_error = 0,
  v_error = 1,
  prior = 0
)
}
\arguments{
\item{n}{number of units}

\item{each}{number of observation per unit}

\item{b0}{intercept}

\item{bj}{a list of coefficients}

\item{distj}{a list of distributions matching bj by position}

\item{paramsj}{a list of parameter vectors matching distj by position}

\item{m_error}{mean of error terms}

\item{v_error}{variance of error terms}

\item{prior}{indicates whether df will serve as main data (0) or prior data (1)}
}
\value{
a data set with an ID for each unit, Y as the dependent variable and X1:Xj
independent variables
}
\description{
Simulate from a Linear Model
}
\examples{
main_lm_data <- sim_lm()
}
