% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lm_dfs.R
\name{sim_lm_dfs}
\alias{sim_lm_dfs}
\title{Simulate Datasets from a Linear Model}
\usage{
sim_lm_dfs(
  n_m = 25,
  b0_m = list(b0 = 0),
  bj_m = list(b1 = 1, b2 = 1),
  distj_m = list(rnorm, rnorm),
  paramsj_m = list(c(0, 1), c(0, 1)),
  transj_m = NULL,
  transrhs_m = NULL,
  intj_m = NULL,
  diste_m = rnorm,
  paramse_m = c(0, 1),
  n_p = 25,
  b0_p = list(b0 = 0),
  bj_p = list(b1 = 1, b2 = 1),
  distj_p = list(rnorm, rnorm),
  paramsj_p = list(c(0, 1), c(0, 1)),
  transj_p = NULL,
  transrhs_p = NULL,
  intj_p = NULL,
  diste_p = rnorm,
  paramse_p = c(0, 1),
  with_prior = 1,
  frames = 2,
  mod_name = list("mod1")
)
}
\arguments{
\item{n_m}{number of units (main data)}

\item{b0_m}{intercept (main data)}

\item{bj_m}{a list of coefficients (main data)}

\item{distj_m}{a list of distributions matching bj_m by position (main data)}

\item{paramsj_m}{a list of parameter vectors matching distj_m by position (main data)}

\item{transj_m}{a list of transformations (infix transformations must specify transrhs_m) (main data)}

\item{transrhs_m}{a list of right-hand-sides for transj_m (set to NULL if transj_m has no rhs) (main data)}

\item{intj_m}{a nested list specifying an interaction between two variable.
The list should take the form list(int1 = list(coef = 1, v1 = "X1", v2 = "X2")) (main data)}

\item{diste_m}{distribution for error term (main data)}

\item{paramse_m}{parameters for the error terms distribution (main data)}

\item{n_p}{number of units (previous data)}

\item{b0_p}{intercept (previous data)}

\item{bj_p}{a list of coefficients (previous data)}

\item{distj_p}{a list of distributions matching bj_p by position (previous data)}

\item{paramsj_p}{a list of parameter vectors matching distj_p by position (previous data)}

\item{transj_p}{a list of transformations (infix transformations must specify transrhs_p) (previous data)}

\item{transrhs_p}{a list of right-hand-sides for transj_p (set to NULL if transj_p has no rhs) (previous data)}

\item{intj_p}{a nested list specifying an interaction between two variable.
The list should take the form list(int1 = list(coef = 1, v1 = "X1", v2 = "X2")) (previous data)}

\item{diste_p}{distribution for error term (previous data)}

\item{paramse_p}{parameters for the error terms distribution (previous data)}

\item{with_prior}{if with_prior = 1, generate data frames that contain main and previous data}

\item{frames}{The number of data frames to generate}

\item{mod_name}{A list of model aliases}
}
\value{
A nested list of data frames (see sim_lm for more information)
}
\description{
Simulate Datasets from a Linear Model
}
\examples{
five_frames <- sim_lm_dfs()
}
