% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lm_run.R
\name{sim_lm_run}
\alias{sim_lm_run}
\title{sim_lm_run}
\usage{
sim_lm_run(df_list, ...)
}
\arguments{
\item{df_list}{a list of data frames built from sim_lm_dfs}

\item{...}{pass parameters to brm function}
}
\value{
a list of of objects containing brm model fits and model generating information
}
\description{
Run brm linear models on list of data frames built with sim_lm_dfs
}
\examples{
linear_mod_dfs <- sim_lm_dfs()
linear_mod_fits <- sim_lm_run(df_list = list(linear_mod_dfs), iter = 3)
}
\references{
Paul-Christian Bürkner (2017). brms: An R Package for Bayesian Multilevel
Models Using Stan. Journal of Statistical Software, 80(1), 1-28.
doi:10.18637/jss.v080.i01
}
