% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sglm.R
\name{sim_sglm}
\alias{sim_sglm}
\title{Simulate Data Using simglm's sim_args Parameter}
\usage{
sim_sglm(data = NULL, sim_args, prior = 0, mod_name = "mod", ...)
}
\arguments{
\item{data}{this parameter was taken from simglm:
https://github.com/lebebr01/simglm/blob/main/R/fixef_sim.r
"Data simulated from other functions to pass to this function.
Can pass NULL if first in simulation string."}

\item{sim_args}{Define model using simglm's sim_args parameter}

\item{prior}{indicates whether df will serve as main data (0) or
prior data (1)}

\item{mod_name}{model alias}

\item{...}{Other arguments to pass to simglm's simulate_fixed,
simulate_randomeffect, simulate_error, or generate_response}
}
\value{
A list with the dataframe produced by the model specification
(as an object named df_mod) and the model specification information
(as an object named df_infor)
}
\description{
Simulate Data Using simglm's sim_args Parameter
}
