% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sglm_comp.R
\name{sim_sglm_comp}
\alias{sim_sglm_comp}
\title{sim_sglm_comp}
\usage{
sim_sglm_comp(fit_obj, with_prior = 1)
}
\arguments{
\item{fit_obj}{Model fits built from the sim_sglm_run function}

\item{with_prior}{Set to 0 if data does not contain prior data.}
}
\value{
A nested list of performance metric data frames.
}
\description{
Compare model performance metrics for sim_sglm_run model fits using measures
of estimator performance found in the simhelpers' calc_absolute function.
}
\references{
Joshi M, Pustejovsky J (2022). _simhelpers: Helper Functions for Simulation
Studies_. R package version 0.1.2,
<https://CRAN.R-project.org/package=simhelpers>.
}
