% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sglm_dfs.R
\name{sim_sglm_dfs}
\alias{sim_sglm_dfs}
\title{sim_sglm_dfs}
\usage{
sim_sglm_dfs(
  data_m = NULL,
  sim_args_m,
  data_p = NULL,
  sim_args_p,
  with_prior = 1,
  frames = 2,
  mod_name = list("mod1"),
  ...
)
}
\arguments{
\item{data_m}{this parameter was taken from simglm:
https://github.com/lebebr01/simglm/blob/main/R/fixef_sim.r
"Data simulated from other functions to pass to this function.
Can pass NULL if first in simulation string." (main data)}

\item{sim_args_m}{Define model using simglm's sim_args parameter (main data)}

\item{data_p}{"Data simulated from other functions to pass to this function.
Can pass NULL if first in simulation string." (previous data)}

\item{sim_args_p}{Define model using simglm's sim_args parameter (previous data)}

\item{with_prior}{if with_prior = 1, generate data frames that contain main and previous data}

\item{frames}{The number of data frames to generate}

\item{mod_name}{A list of model aliases}

\item{...}{Other arguments to pass to simglm's simulate_fixed,
simulate_randomeffect, simulate_error, or generate_response}
}
\value{
A nested list of data frames (see sim_sglm for more information)
}
\description{
Simulate main and previous datasets using simglm's sim_args parameter
}
\references{
LeBeau B (2022). _simglm: Simulate Models Based on the Generalized Linear
Model_. R package version 0.8.9, <https://CRAN.R-project.org/package=simglm>.
}
