% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sglm_run.R
\name{sim_sglm_run}
\alias{sim_sglm_run}
\title{Run brm simglm Models on List of Data Frames}
\usage{
sim_sglm_run(df_list, ...)
}
\arguments{
\item{df_list}{a list of data frames built from sim_sglm_dfs}

\item{...}{pass parameters to brm function}
}
\value{
a list of of objects containing brm model fits and model generating information
}
\description{
Run brm simglm Models on List of Data Frames
}
