% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lm.R
\name{sim_lm}
\alias{sim_lm}
\title{Simulate Data from a Linear Model}
\usage{
sim_lm(
  n = 25,
  b0 = list(b0 = 0),
  bj = list(b1 = 1, b2 = 1),
  distj = list(rnorm, rnorm),
  paramsj = list(c(0, 0), c(1, 1)),
  transj = list(`^`, `^`),
  transrhs = list(1, 1),
  intj = NULL,
  diste = rnorm,
  paramse = c(0, 1),
  prior = 0,
  mod_name = "mod"
)
}
\arguments{
\item{n}{number of units}

\item{b0}{intercept}

\item{bj}{a list of coefficients}

\item{distj}{a list of distributions matching bj by position}

\item{paramsj}{a list of parameter vectors matching distj by position}

\item{transj}{a list of transformations (infix transformations must specify transrhs)}

\item{transrhs}{a list of right-hand-sides for transj (set to NULL if transj has no rhs)}

\item{intj}{a nested list specifying an interaction between two variable.
The list should take the form list(int1 = list(coef = 1, v1 = "X1", v2 = "X2"))
to specify an interaction between X1 and X2 with a regression coefficient of 1.}

\item{diste}{distribution for error term}

\item{paramse}{parameters for the error terms distribution}

\item{prior}{indicates whether df will serve as main data (0) or prior data (1)}

\item{mod_name}{model alias}
}
\value{
A list object.  First element is a data set with an ID for each unit, Y as the dependent variable and X1:Xj
independent variables (plus interactions if specified).  Second element is the
model alias and input parameters (b0, b1:bj, prior indicator, plus interactions).
}
\description{
Simulate Data from a Linear Model
}
\examples{
#Build df using default parameters
main_lm_data <- sim_lm()

#Build df with an interaction
main_lm_data_in <- sim_lm(intj = list(int1 = list(coef = 5, v1 = "X1", v2 = "X2")))
}
