% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lm_comp.R
\name{sim_lm_comp}
\alias{sim_lm_comp}
\title{Compare Linear Model Performance Metrics}
\usage{
sim_lm_comp(fit_obj)
}
\arguments{
\item{fit_obj}{A list of model fits built from the sim_lm_run function}
}
\value{
a nested list of performance metric data frames.  One top level list for the main data
and one top level list for the prior data.  Within each top level list, there is one data frame of
performance metrics for each model.
}
\description{
Compare Linear Model Performance Metrics
}
