% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lm_comp.R
\name{sim_lm_comp}
\alias{sim_lm_comp}
\title{sim_lm_comp}
\usage{
sim_lm_comp(fit_obj, with_prior = 1)
}
\arguments{
\item{fit_obj}{A list of model fits built from the sim_lm_run function}

\item{with_prior}{Set to 0 if data does not contain prior data.}
}
\value{
a nested list of performance metric data frames.  One top level list for the main data
and one top level list for the prior data.  Within each top level list, there is one data frame of
performance metrics for each model.
}
\description{
Compare linear model performance metrics from sim_lm_run model fits using
measures of estimator performance found in the simhelpers' calc_absolute
function.
}
\references{
Joshi M, Pustejovsky J (2022). _simhelpers: Helper Functions for Simulation
Studies_. R package version 0.1.2,
<https://CRAN.R-project.org/package=simhelpers>.
}
