% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lm_run.R
\name{sim_lm_run}
\alias{sim_lm_run}
\title{Run brm Linear Models on List of Data Frames}
\usage{
sim_lm_run(df_list, ...)
}
\arguments{
\item{df_list}{a list of data frames built from sim_lm_dfs}

\item{...}{pass parameters to brm function}
}
\value{
a list of of objects containing brm model fits and model generating information
}
\description{
Run brm Linear Models on List of Data Frames
}
\examples{
linear_mod_dfs <- sim_lm_dfs()
linear_mod_fits <- sim_lm_run(df_list = list(linear_mod_dfs), iter = 3)
}
