% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sglm_comp.R
\name{sim_sglm_comp}
\alias{sim_sglm_comp}
\title{Compare Model Performance Metrics}
\usage{
sim_sglm_comp(fit_obj, with_prior = 1)
}
\arguments{
\item{fit_obj}{Model fits built from the sim_sglm_run function}

\item{with_prior}{Set to 0 if data does not contain prior data.}
}
\value{
A nested list of performance metric data frames.  One top level list
for the main data and one top level list for the prior data.  Within each top
level list, there is one data frame of performance metrics for each
}
\description{
Compare Model Performance Metrics
}
